Separate out vulkan setup from litmus generator
